feat: Add list-printers.js to get a list of printers
This commit adds a new file, `list-printers.js`, which uses the `pdf-to-printer` library to get a list of available printers and logs them to the console. This feature will be useful for users who want to select a specific printer for printing PDF documents.

The `list-printers.js` file is added to the project, and the necessary dependencies are imported and used to fetch the list of printers. Any errors that occur during the process are also logged to the console.

This commit also includes some changes to the `.gitignore` file to exclude the `/dist_electron` directory, which will be used for Electron builds. Additionally, the `vue.config.js` file is updated to configure the Electron Builder options for the project.

Lastly, the `package.json` file is modified to include new scripts for building and serving the Electron app, as well as post-install and post-uninstall scripts for installing app dependencies. The `main` field is also added to specify the entry point for the Electron app.

These changes are necessary to support the upcoming Electron integration in the project and provide the ability to list available printers for printing PDF documents.